Add OBB support to TFLiteDetector
Resolves #8

I've also taken this as an opportunity to refactor the way post
processing is handled, the way it was done before seemed quite clunky
and hard to follow. This should be easier to use and requires less
duplicated logic.
